+++ includes/gateways/class-wpsms-gateway-africastalking.php
@@ -10,6 +10,7 @@ class africastalking extends \WP_SMS\Gateway
     public $unit;
     public $flash = "false";
     public $isflash = false;
+    public $supportIncoming = true;
 
     public function __construct()
     {