+++ wp-sms.php
@@ -2,8 +2,8 @@
 /**
  * Plugin Name: WP SMS
  * Plugin URI: https://wp-sms-pro.com/
- * Description: The Best WordPress SMS Messaging and Notification Plugin for WordPress!
- * Version: 6.9
+ * Description: Ultimate SMS & MMS Notifications, 2FA, OTP, and Integrations with WooCommerce, GravityForms, and More
+ * Version: 6.10
  * Author: VeronaLabs
  * Author URI: https://veronalabs.com/
  * Text Domain: wp-sms